Move schema functions to new file
Split schema functions into smaller units
