+++ Cargo.toml
@@ -12,7 +12,7 @@ crate-type = ["cdylib"]
 name = "configcrunch"
 
 [dependencies]
-pyo3 = { version = "0.27", features = ["extension-module"] }
+pyo3 = { version = "0.27", features = ["extension-module", "multiple-pymethods"] }
 serde = { version = "1", features = ["derive"] }
 serde_yaml = "0.9"
 minijinja = "2"
